Breaking complex class into traits and adding method param/return types, correcting doc block definitions
