+++ composer.json
@@ -17,6 +17,7 @@
     ],
     "require": {
         "php": ">=7.2",
+        "ext-simplexml": "*",
         "j4mie/paris": "^1.5",
         "firebase/php-jwt": "^5.0",
         "gumlet/php-image-resize": "^1.9",
@@ -29,7 +30,7 @@
         "phpmailer/phpmailer": "^6.0",
         "twig/twig": "^2.7",
         "php-curl-class/php-curl-class": "^8.6",
-        "gerardbalaoro/figlet-php": "^1.0"
+        "gerardbalaoro/figlet-php": "^1.0",
     },
     "require-dev": {
         "phpunit/phpunit": "^8.0",
