+++ tests/unit/Http/ResponseTest.php
@@ -5,9 +5,22 @@ namespace Quantum\Test\Unit;
 use PHPUnit\Framework\TestCase;
 use Quantum\Exceptions\StopExecutionException;
 use Quantum\Http\Response;
+use Quantum\Libraries\Storage\FileSystem;
+use Quantum\Loader\Loader;
 
 class ResponseTest extends TestCase
 {
+
+    public function setUp(): void
+    {
+        Response::init();
+
+        $loader = new Loader(new FileSystem);
+
+        $loader->loadDir(dirname(__DIR__, 3) . DS . 'src' . DS . 'Helpers' . DS . 'functions');
+
+        $loader->loadFile(dirname(__DIR__, 3) . DS . 'src' . DS . 'constants.php');
+    }
 
     public function tearDown(): void
     {
